The bind command causes x2go to not start anymore
I'm using x2go for remote access to a server. 
It seems that when starting the x2go desktop a shell is started without creating a pty. 

This causes the bind command to fail, and even though this should only a warning x2go treats it as an error.
The error is "bash bind warning: line editing not enabled" 
Checking whether the completion script is run in a terminal  solves this issue, see also
https://superuser.com/questions/892658/remote-ssh-commands-bash-bind-warning-line-editing-not-enabled